Delete ExampleUnitTest.kt. Rename test specific util directory from util to testutil.
